Switched provisioning mats from "Tier #" to "Flavor" labels
Based on feedback from vestig at esoui, I'm changing the labels for all of the "Tier 2" and "Tier 3" ingredients, to something more descriptive. Tier 2 becomes "For extra flavor in blue and purple recipes", and Tier 3 becomes "For extra flavor in purple recipes".

I've also re-sorted the table of provisioning materials so it's easier to find items when looking at the ProvisioningMaterials.lua file.